requirements.txt
python-docx==1.1.2
Pillow==11.2.1
PyYAML>=6.0
